refactor(onboard): decompose viewer god-object + one authoritative frontend (#461 phase 2)
Findings #2 and #4.

#2 server.go (1814 lines, ≥7 responsibilities) decomposed into cohesive
files within the package:
  - server.go    — HTTP plumbing only (routing, static UI, writeJSON)
  - store.go     — RecordingStore: the replaydata filesystem repository
  - models.go    — the response DTOs (previously scattered in handlers)
  - scenarios.go — /api/scenarios list + detail + helpers
  - recordings.go — archived-recording browsing
  - catalog.go   — coverage matrix + annotation passes
  - recipe.go    — recipe catalog + coverage_id dedup
  - spec.go      — scenario-spec markdown parsing

Introduces a RecordingStore seam: handlers ask it for scenario files,
scenario/archive listings, and expected.jsonl presence instead of
hand-building paths with inline os.ReadFile/os.ReadDir. The /api/catalog
3-stage annotation pipeline, which unmarshalled+marshalled the whole
catalog JSON three times, now parses once / annotates in place / marshals
once. Dead resolveScenarioFolderFromMap removed.

#4 One authoritative frontend. platforms/web/ is the live session
dashboard (daemon + viewer); the viewer's embedded internal/viewer/web/
SPA is the separate catalog/scenario browser. Documented in AGENTS.md and
the server.go package doc. The brittle exact-string </head> splice that
injected the ws-diag script is replaced by injectBeforeClosingTag —
case-insensitive, falls back to </body>, logs + appends rather than
silently dropping the script if neither tag exists.

Tests: TestInjectBeforeClosingTag covers the casing/fallback/append cases.
Existing server/playback tests pass against the decomposed package.
